// packages/shared/src/schemas/domain.ts
export const timestampSchema = z.union([z.date(), z.string(), z.number()])
// ...
export const platformPreferenceProfileSchema = z.object({
  userId: z.string(),
  organizationId: z.string(),
  platform: z.enum(platforms),
  theme: z.enum(["light", "dark", "system"]),
  accentColor: z.string(),
  boldText: z.boolean(),
  showTips: z.boolean().default(true),
  createdAt: timestampSchema,
  updatedAt: timestampSchema
})